Added CalculateOdds function
CalculateOdds recursively fills the board for each opponent hand and adds upp all the wins, losses and draws. As Testing for this is somewhat difficult, I used known odds for know hands to test the function, which is why I used the normalized odds. The total number of combinations is correct, however the win percentages are a bit off. For now the function works with 1.5% accuracy.