Refactoring code to use only a single channel. Limit number of go routines created #2
As per #2 I refactored code so that we use only one channel and make a go routine only for each opponent hand, not all possible board combinations. As expected when processing only a single opponent hand the processing time returns to double what we currently have, but when processing multiple hands the speed is the same.